% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boa.R
\name{boa}
\alias{boa}
\title{Implementing the BOA algorithm with R}
\usage{
boa(data, classes, prior = "BetaBinomial", method = "pattern", params,
  iter_max = 100, cool_rate = 10, p = 0.1)
}
\arguments{
\item{data}{a data frame containing the variables for the model. No continuous data is allowed.}

\item{classes}{a vector of classes (0 or 1) for the data.}

\item{prior}{prior: "BetaBinomial" or "Poisson"}

\item{method}{method used for finding the minima: either "pattern" or "literal"}

\item{params}{list of parameters for the model based on the choice of prior}

\item{iter_max}{maximum number of iterations}

\item{cool_rate}{cooling rate}

\item{p}{prob of selecting a random neighbor}
}
\value{
vector of classes
}
\description{
Implementing the BOA algorithm with R
}
\examples{
params <- list(max_length = 3, alpha = rep(10, 3), beta = rep(10, 3),
         alpha_plus = 500, alpha_minus = 500, beta_plus = 1, beta_minus = 1)
df <- TicTacToe[, -28]
dfClasses <- TicTacToe$outcome
b <- boa(data = df, classes = dfClasses, prior = "BetaBinomial", 
         method = "pattern", params, iter_max = 50, cool_rate = 1000, 
         p = 0.1)
b
}
