% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_less.R
\name{.cover_less}
\alias{.cover_less}
\title{Auxiliary function to change the pattern set}
\usage{
.cover_less(pattern_set, method, p, data, classes, prior, params, 
        pattern_pool)
}
\arguments{
\item{pattern_set}{the pattern set}

\item{method}{method used for finding the minima: either "pattern" or "literal"}

\item{p}{the probability of returning a random neighbor (as opposed to the one
with the minimum score function)}

\item{data}{a data frame containing the variables for the model.
No continuous data is allowed.}

\item{classes}{a vector of classes (0 or 1) for the data.}

\item{prior}{prior: "BetaBinomial" or "Poisson"}

\item{params}{list of parameters for the model based on the choice of prior}

\item{pattern_pool}{the set of all possible patterns for \code{pattern_set} mined from the data}
}
\value{
a new pattern set
}
\description{
\code{.cover_less()} returns one of the neighbors of \code{pattern_set} that
results in less positively classified observations. How neighbor is defined
depends on the parameter \code{method}
}
\keyword{internal}
