% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_more.R
\name{.cover_more}
\alias{.cover_more}
\title{Auxiliary function to change the pattern set \code{pattern_set}}
\usage{
.cover_more(pattern_set, method, p, data, classes, prior, params, pattern_pool)
}
\arguments{
\item{pattern_set}{the pattern set}

\item{method}{method used for finding the minima: either "pattern" or "literal"}

\item{p}{the probability of returning a random neighbor (as opposed to the one
with the minimum score function)}

\item{data}{a data frame containing the variables for the model.
No continuous data is allowed.}

\item{classes}{a vector of classes (0 or 1) for the data.}

\item{prior}{prior: "BetaBinomial" or "Poisson"}

\item{params}{list of parameters for the model based on the choice of prior}

\item{pattern_pool}{the set of all possible patterns for \code{pattern_set} mined from the data}
}
\value{
a new pattern set
}
\description{
'.cover_more()' returns one of the neighbors of \code{pattern_set} that
results in more positively classified observations
}
\keyword{internal}
