% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{.score}
\alias{.score}
\title{Calculates the score function}
\usage{
.score(data, classes, pattern_set, prior, params, pattern_pool)
}
\arguments{
\item{data}{a data frame containing the variables for the model.
No continuous data is allowed.}

\item{classes}{a vector of classes (0 or 1) for the data.}

\item{pattern_set}{the pattern set}

\item{prior}{prior: "BetaBinomial" or "Poisson"}

\item{params}{list of parameters for the model based on the choice of prior}

\item{pattern_pool}{the set of all possible patterns for \code{pattern_set} mined from the data}
}
\value{
score
}
\description{
'.score()' is an auxiliary function to calculate the score function
(defined as negative logarithm of the likelihood) for the given set of patterns
}
\keyword{internal}
