% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_patterns.R
\name{.screen_patterns}
\alias{.screen_patterns}
\title{Auxiliary function screenPatterns (using information gain)}
\usage{
.screen_patterns(data, classes, pattern_pool, num_pat_max = 2000)
}
\arguments{
\item{data}{a data frame containing the variables in the model}

\item{classes}{a vector of classes (0 or 1) for the data.}

\item{pattern_pool}{the set of all possible patterns for \code{pattern_set} mined from the data}

\item{num_pat_max}{how many patterns to keep}
}
\value{
a list of all patterns in the given data, indexed by lengths
}
\description{
The function screens for the available patterns to pick best ones based
on the information gain
}
\keyword{internal}
