% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.boa.R
\name{predict.boa}
\alias{predict.boa}
\title{Prediction function}
\usage{
\method{predict}{boa}(object, data, ...)
}
\arguments{
\item{object}{pattern set}

\item{data}{data frame containing the variables for the model.
No continuous data is allowed.}

\item{...}{additional arguments for specific methods.}
}
\value{
a vector of predictions
}
\description{
Prediction function
}
