% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOAR.R
\docType{package}
\name{SOAR}
\alias{SOAR}
\title{SOAR}
\description{
Implementing Or's of And's algorithm
}
\references{
\itemize{
  \item Wang, T., Rudin, C., Doshi-Velez, F., Liu Y., Klampfl, E. & MacNeille E. (2015).
        Bayesian Or’s of And’s for Interpretable Classification with Application
        to Context Aware Recommender Systems. Technical report, MIT.
        \url{https://pdfs.semanticscholar.org/ee43/4531203bf7609c9648f78eacf4b7ea750dd0.pdf}.
  \item Borgelt, C. (2005). An Implementation of the FP-growth Algorithm. ACM.
  }
}
\author{
Elena Khusainova <elena.r.khusainova@gmail.com>

Ritwik Mitra

Emily Dodwell
}
