% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicTacToe.R
\docType{data}
\name{TicTacToe}
\alias{TicTacToe}
\title{TicTacToe data}
\format{An object of class \code{data.frame} with 958 rows and 28 columns.}
\usage{
TicTacToe
}
\description{
TicTacToe data
}
\keyword{datasets}
